rename datatableproviders to data-table-providers
