+++ src/pages/workspaces/workspace/workflows/DataStepContent.js
@@ -8,7 +8,7 @@ import DataTable from 'src/components/data/DataTable'
 import { icon } from 'src/components/icons'
 import { ValidatedInput } from 'src/components/input'
 import Modal from 'src/components/Modal'
-import { EntityServiceDataTableProvider } from 'src/libs/ajax/datatableproviders/EntityServiceDataTableProvider'
+import { EntityServiceDataTableProvider } from 'src/libs/ajax/data-table-providers/EntityServiceDataTableProvider'
 import { FormLabel } from 'src/libs/forms'
 import * as Style from 'src/libs/style'
 import * as Utils from 'src/libs/utils'
